fix: move feature_engineering into package, fix relative import (#232)
feature_engineering.py lived in the top-level pipeline/model/ tree which
is not an installable package. train_model.py imported it with a relative
path (from pipeline.model...) that only resolves when the working directory
is the repo root, breaking cloud runs via arm_orbit and any invocation from
a different directory.

Move feature_engineering.py into the installable package at
asf_heat_pump_suitability/pipeline/model/block_of_flats/ (consistent with
the SPEC convention that support modules live inside the package) and update
the import to use the full absolute package path.

Closes #232
